% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leeds_box.R
\name{leeds_box}
\alias{leeds_box}
\title{leeds boxplot}
\usage{
leeds_box(colour, seed)
}
\arguments{
\item{colour}{string for colour, one of "yellow",
"blue" or "anniversary_gold"}

\item{seed}{number to set the seed for the random sample}
}
\value{
a list with 6 components, see ?boxplot
}
\description{
Displays a boxplot for 100 random numbers
from a standard normal distribution in one of three colours
(Leeds blue, Leeds yellow or Leeds anniversary gold)
specified by the user
}
\examples{
leeds_box("anniversary_gold", 4)
}
